refactor: atualizar imports para usar alias
- Alterados imports de mocks para utilizar alias '@/mocks'
- Ajustados arquivos de teste e serviços para refletir nova estrutura
